refactor: split models and types
